implementación de selección de casillas

+++ Jaque-bot/V2D.h
@@ -2,8 +2,8 @@
 class V2D
 {
 public:
-	V2D(int xv = 0, int yv = 0) : x(xv), y(yv) {};
-	int x, y;
+	V2D(float xv = 0, float yv = 0) : x(xv), y(yv) {};
+	float x, y;
 
 	V2D operator - (V2D v);
 };
